seperate mcFilter logic into lib file
